refactor(settings): split into subroutes; preserve toasts and CRUD
