+++ src/pages/settings/BrandsPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const BrandsPage: React.FC = () => {
+  return (
+    <div className="bg-white p-6 rounded-lg shadow">
+      <h2 className="text-2xl font-bold text-gray-800">Manage Brands</h2>
+      <p className="mt-2 text-gray-600">Configuration for this section will be available here.</p>
+    </div>
+  );
+};
+
+export default BrandsPage;
